requirements.txt
certifi==2023.7.22
charset-normalizer==3.3.2
idna==3.4
loguru==0.7.2
python-dotenv==1.0.0
pytube==15.0.0
requests==2.31.0
schedule==1.2.1
scrapetube==2.5.1
typing_extensions==4.8.0
urllib3==2.0.7
